+++ CHANGELOG.md
@@ -55,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Revert zoomed in view [#99](https://github.com/pLTdwaxn/Caroushow/issues/99)
 - Bump Expo version [#102](https://github.com/pLTdwaxn/Caroushow/issues/102)
 - Disabled runCropper on parameters change [#100](https://github.com/pLTdwaxn/Caroushow/issues/100)
+- Move ImagePreview to top of the screen [#109](https://github.com/pLTdwaxn/Caroushow/issues/109)
 
 ### Removed